Added licenses, javadoc, fixed a few formatting issues and moved AbstractFileServerToBagJob into the normalization package
